resolve reordering tweets bug
